Crate riddle_image

Riddle crate for loading and manipulating image data in main memory.

Built largely on the back of ::image and its dependencies.

Example

// Load an image from a png
let png_bytes = include_bytes!("../../example_assets/image.png");
let png_img = Image::load(&png_bytes[..], ImageFormat::Png)?;

// Make a blank image and blit the png on to it
let mut blank_img = Image::new(256, 256);
blank_img.blit(&png_img, [0, 0].into());

Modules

image_ext
Traits which expose underlying image crate's types

Structs

Color
An RGBA color, parameterized over color channel type.

Image
A representation of an image stored in main memory. The image is stored as RGBA32.

Enums

ImageError

ImageFormat
The set of support image file formats which Image can load