Function riddle_image::filters::distance_field

pub fn distance_field(source: Image, scale: f64) -> Image

Generate a distance field for the given image using 8SSEDT. The scale argument roughly maps to how many pixels deep the boundary region is between min and max values.

Examples

let png_bytes = include_bytes!("../../examples/image-distance-field/sample.png");
let png_img = Image::load(&png_bytes[..], ImageFormat::Png)?;

let processed_img = filters::distance_field(png_img, 10.0);

Resources:

http://www.codersnotes.com/notes/signed-distance-fields/

https://github.com/Lisapple/8SSEDT